# Reduce JSON deserialization calls, add benchmarks