--- a/benchmarks/ConfigCatClient.Benchmarks.csproj
+++ b/benchmarks/ConfigCatClient.Benchmarks.csproj
@@ -1,0 +1,34 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <Configurations>Debug;Release;Benchmark</Configurations>
+  </PropertyGroup>
+
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Benchmark|AnyCPU'">
+    <Optimize>true</Optimize>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
+    <PackageReference Include="ConfigCat.Client" Version="6.1.20" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Reference Include="ConfigCat.Client.Benchmark">
+      <HintPath>..\src\ConfigCatClient\bin\Benchmark\netstandard2.1\ConfigCat.Client.Benchmark.dll</HintPath>
+      <Aliases>from_project</Aliases>
+      <Private>true</Private>
+    </Reference>
+  </ItemGroup>
+
+  <Target Name="ChangeAliasOfNugetReferencedStashbox" BeforeTargets="FindReferenceAssembliesForReferences;ResolveReferences">
+    <ItemGroup>
+      <ReferencePath Condition="'%(FileName)' == 'ConfigCat.Client'">
+        <Aliases>from_nuget</Aliases>
+      </ReferencePath>
+    </ItemGroup>
+  </Target>
+
+</Project>