--- a/benchmarks/Program.cs
+++ b/benchmarks/Program.cs
@@ -1,0 +1,15 @@
+﻿using BenchmarkDotNet.Running;
+using System;
+
+namespace ConfigCatClient.Benchmarks
+{
+    class Program
+    {
+        private static void Main(string[] args)
+        {
+            BenchmarkRunner.Run<GetValueBenchmarks>();
+
+            Console.ReadKey();
+        }
+    }
+}