--- a/src/ConfigCatClient/ConfigCatClient.csproj
+++ b/src/ConfigCatClient/ConfigCatClient.csproj
@@ -81,12 +81,20 @@ Works with .NET, .NET Core, .NET Standard</Description>
           <NeutralLanguage></NeutralLanguage>
           <DocumentationFile>bin\$(Configuration)\$(TargetFramework)\ConfigCatClient.xml</DocumentationFile>
           <DebugType>portable</DebugType>
+          <Configurations>Debug;Release;Benchmark</Configurations>
      </PropertyGroup>
 
      <ItemGroup Condition=" '$(TargetFramework)' == 'net45' ">
           <Reference Include="System" />
           <Reference Include="Microsoft.CSharp" />
      </ItemGroup>
+
+    <PropertyGroup Condition="'$(Configuration)' == 'Benchmark'">
+        <AssemblyName>ConfigCat.Client.Benchmark</AssemblyName>
+        <AssemblyTitle>ConfigCat.Client.Benchmark</AssemblyTitle>
+        <RootNamespace>ConfigCat.Client.Benchmark</RootNamespace>
+        <Optimize>true</Optimize>
+    </PropertyGroup>
 
      <ItemGroup>
           <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />