### Describe the purpose of your pull request

Reducing the number of JSON deserialization calls by caching the last deserialized object in memory. Now, the deserialization is called only when the retrieved JSON string differs from the latest cached one.

Performance benchmark:
```
|      Method |       Mean |     Error |    StdDev | Ratio |  Gen 0 |  Gen 1 | Allocated |
|------------ |-----------:|----------:|----------:|------:|-------:|-------:|----------:|
| OldGetValue | 109.117 us | 0.2605 us | 0.2436 us |  1.00 | 2.9297 | 0.1221 |     25 KB |
| NewGetValue |   7.404 us | 0.0256 us | 0.0214 us |  0.07 | 0.9308 |      - |      8 KB |
```

### Related issues (only if applicable)

#22 

### Requirement checklist (only if applicable)

- [ ] I have covered the applied changes with automated tests.
- [x] I have executed the full automated test set against my changes.
- [x] I have validated my changes against all supported platform versions.