@z4kn4fein 2021-09-16
### Update
Replaced `.GetHashCode()` with `MurmurHash3`. Far less chance for hashes to collide.

Updated benchmark:
```
| Method |       Mean |     Error |    StdDev | Ratio |  Gen 0 |  Gen 1 | Allocated |
|------- |-----------:|----------:|----------:|------:|-------:|-------:|----------:|
|    Old | 107.847 us | 0.4968 us | 0.4647 us |  1.00 | 2.9297 | 0.1221 |     25 KB |
|    New |   6.921 us | 0.0347 us | 0.0325 us |  0.06 | 1.3809 | 0.0076 |     11 KB |
```